use CSV reader in simulation class
this no longer requires holding all data in memory, instead the next
line is read the next call to getAxes happens. The timestamps in the
file are simply ignored.

The whole thing will fail once all values are read from the file, but it
should be large enough to last a couple of hours - good enough for
testing